src/config/constant.py
import os

from dotenv import load_dotenv

load_dotenv()

class geminiAiCFG:
    API_KEY = os.environ["GEMINI_API_KEY"]
    API_MODEL = os.getenv("GEMINI_API_MODEL", "gemini-pro")

class zillizCFG:
    ZILLIZDB_USERNAME = os.environ["ZILLIZDB_USERNAME"]
    ZILLIZDB_PASSWORD = os.environ["ZILLIZDB_PASSWORD"]
    ZILLIZDB_HOST = os.environ["ZILLIZDB_HOST"]
    ZILLIZDB_PORT = os.environ["ZILLIZDB_PORT"]
    ZILLIZDB_COLLECTION_NAME = os.environ["ZILLIZDB_COLLECTION_NAME"]
    CLIP_MODEL_NAME = os.environ["CLIP_MODEL_NAME"]

