requirements.txt
asgi_correlation_id==4.2.0
fastapi==0.110.1
pydantic==2.6.1
pytest==7.3.2
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
tenacity==8.2.3
neo4j==5.17.0
pillow==10.2.0
pymongo==4.6.2
sentence-transformers==2.5.1
matplotlib==3.8.3
opencv-python==4.9.0.80
gunicorn==21.2.0
pandas==2.2.1